requirements.txt
kafka-python-ng
psycopg2-binary
Faker
yfinance
pandas
numpy
streamlit
